requirements.txt
asgiref==3.3.4
Django==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
gunicorn==20.0.4
pymysql == 0.7.4
mysqlclient == 2.0.3
django-cors-headers == 3.7.0